# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCQ Online Test Platform API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/examinee/sessions/{session_id}/questions/{position}": {
      "get": {
        "tags": [
          "examinee"
        ],
        "summary": "Get question at position",
        "description": "Fetch the question at a given position (1-50) in the examinee's session.\n\nThe correct answer is never included. The response also returns the\nexaminee's previously selected option for that question, if any.",
        "operationId": "get_question_at_position_api_v1_examinee_sessions__session_id__questions__position__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "position",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The question at the requested position, with any previously selected option.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ExamineeQuestionAtPositionResource_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/examiner/questions": {
      "get": {
        "tags": [
          "examiner-questions"
        ],
        "summary": "List questions",
        "description": "List question bank entries with cursor-based pagination.\n\nIncludes the correct answer for each question (examiner view) and the\ncurrent `active_count` of non-deleted questions.",
        "operationId": "list_questions_api_v1_examiner_questions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size (1-100).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size (1-100)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination \u2014 pass the `next_cursor` from the previous page.",
              "title": "Cursor"
            },
            "description": "Cursor for pagination \u2014 pass the `next_cursor` from the previous page."
          },
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include soft-deleted questions in the listing.",
              "default": false,
              "title": "Include Deleted"
            },
            "description": "Include soft-deleted questions in the listing."
          }
        ],
        "responses": {
          "200": {
            "description": "A page of questions plus the total active question count.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_QuestionListData_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "examiner-questions"
        ],
        "summary": "Create a question",
        "description": "Add a multiple-choice question (four options, one correct) to the bank.\n\nRequires an examiner session and the `X-CSRF-Token` header.",
        "operationId": "create_question_api_v1_examiner_questions_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created question.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_QuestionOut_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
    "/api/v1/examiner/questions/{question_id}": {
      "patch": {
        "tags": [
          "examiner-questions"
        ],
        "summary": "Update a question",
        "description": "Partially update a question; omitted fields keep their current values.\n\nEach edit creates a new question version so in-flight test sessions keep\nreferencing the version they were served. Requires an examiner session\nand the `X-CSRF-Token` header.",
        "operationId": "patch_question_api_v1_examiner_questions__question_id__patch",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Question Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated question with an incremented version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_QuestionOut_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "examiner-questions"
        ],
        "summary": "Soft-delete a question",
        "description": "Soft-delete a question (it is excluded from new tests but retained for history).\n\nRejected with 409 if it would drop the active bank below the minimum\nrequired for test generation. Requires an examiner session and the\n`X-CSRF-Token` header.",
        "operationId": "delete_question_api_v1_examiner_questions__question_id__delete",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Question Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Question soft-deleted."
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "409": {
            "description": "Deletion would reduce the active question bank below the required minimum (default 50).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The resource was modified by another request.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuccessResponse_ExamineeQuestionAtPositionResource_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExamineeQuestionAtPositionResource",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExamineeQuestionAtPositionResource]"
      },
...
      "SuccessResponse_QuestionListData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuestionListData",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[QuestionListData]"
      },
      "SuccessResponse_QuestionOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuestionOut",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[QuestionOut]"
      },
...
      "ProblemDetails": {
        "description": "RFC 7807 problem details returned with content type application/problem+json.",
        "properties": {
          "type": {
            "description": "URI identifying the problem type.",
            "examples": [
              "https://api.example.com/problems/unauthenticated"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "examples": [
              "Unauthenticated"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "examples": [
              401
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Human-readable explanation specific to this occurrence.",
            "examples": [
              "Authentication required."
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path on which the problem occurred.",
            "examples": [
              "/api/v1/auth/me"
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request identifier for log correlation; matches X-Request-Id.",
            "examples": [
              "9f1b2c3d4e5f6a7b"
            ],
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Field-level validation errors (only on 400 validation failures).",
            "title": "Errors"
          },
          "locked_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "When a locked account becomes available again (only on 423 responses).",
            "title": "Locked Until"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetails",
        "type": "object"
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Opaque server-side session token issued by the login/entry endpoints as an HttpOnly cookie. Examiner sessions last 1 hour, examinee sessions 4 hours by default."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound CSRF token required on mutating requests. Obtain it from GET /auth/me (or GET /auth/csrf before login)."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}